// types/rawResponseType.ts
export interface RawResponseType {
    graphql: Graphql;
}

export interface Graphql {
    shortcode_media: ShortcodeMedia;
}

export interface ShortcodeMedia {
    __typename:                      string;
    id:                              string;
    shortcode:                       string;
    dimensions:                      Dimensions;
    gating_info:                     null;
    fact_check_overall_rating:       null;
    fact_check_information:          null;
    sensitivity_friction_info:       null;
    sharing_friction_info:           SharingFrictionInfo;
    media_overlay_info:              null;
    media_preview:                   null | string;
    display_url:                     string;
    display_resources:               DisplayResource[];
    is_video:                        boolean;
    tracking_token:                  string;
    edge_media_to_tagged_user:       EdgeMediaToTaggedUser;
    edge_media_to_caption:           EdgeMediaToCaptionClass;
    caption_is_edited:               boolean;
    has_ranked_comments:             boolean;
    edge_media_to_parent_comment:    EdgeMediaToParentCommentClass;
    edge_media_to_hoisted_comment:   EdgeMediaToCaptionClass;
    edge_media_preview_comment:      EdgeMediaPreviewComment;
    comments_disabled:               boolean;
    commenting_disabled_for_viewer:  boolean;
    taken_at_timestamp:              number;
    edge_media_preview_like:         EdgeMediaPreviewLike;
    edge_media_to_sponsor_user:      EdgeMediaToCaptionClass;
    location:                        Location | null;
    viewer_has_liked:                boolean;
    viewer_has_saved:                boolean;
    viewer_has_saved_to_collection:  boolean;
    viewer_in_photo_of_you:          boolean;
    viewer_can_reshare:              boolean;
    owner:                           Owner;
    is_ad:                           boolean;
    edge_web_media_to_related_media: EdgeMediaToCaptionClass;
    edge_sidecar_to_children?:       EdgeSidecarToChildren;
    edge_related_profiles:           EdgeMediaToCaptionClass;
    accessibility_caption?:          null | string;
    dash_info?:                      DashInfo;
    has_audio?:                      boolean;
    video_url?:                      string;
    video_view_count?:               number;
    video_play_count?:               null;
    encoding_status?:                string;
    is_published?:                   boolean;
    product_type?:                   string;
    title?:                          string;
    video_duration?:                 number;
    thumbnail_src?:                  string;
    clips_music_attribution_info?:   null;
}